+++ src/main/java/net/lingala/zip4j/util/InternalZipConstants.java
@@ -35,7 +35,7 @@ public interface InternalZipConstants {
     int ZIP64_ENDSIG_LOC = 0x07064b50;
     int ZIP64_ENDSIG = 0x06064b50;
     static int EXTRAFIELDZIP64LENGTH = 0x0001;
-    short AESSIG = (short)0x9901;
+    int AESSIG = 0x9901;
 
     static final int STD_DEC_HDR_SIZE = 12;
